requirements.txt
streamlit
google-genai
supabase
requests
beautifulsoup4
